Finding the Production Indexes

BDisposal.prodIndex — Function

prodIndex(gI,gO,bO,bI;retToScale,prodStructure,convexAssumption,startθ,startμ,startλ)

Compute productivity indexes

Given a set of measures of inputs, "good" ("desiderable") and "bad" ("undesiderable") outputs for different decision making units, compute their productivity indexes improvements (or declines) between consecutive time periods.

Parameters:

Positional

gI: "Good" inputs (3D matrix by DMUs, input items and periods)

gO: "Good" outputs (3D matrix by DMUs, input items and periods)

bO: "Bad" outputs (3D matrix by DMUs, input items and periods)

bI: "Bad" inputs (optional 3D matrix by DMUs, input items and periods) [default: empty array]

Keyword

retToScale: Wheter the return to scales should be assumed "constant" (default) or "variable"

prodStructure: Wheter the production structure should be assumed "additive" (default) or "multiplicative"

convexAssumption: Wheter a convex production frontier should be assumed [default: true]

startθ,startμ,startλ: Initial values for the convex optimisation problem (used only with convexAssumption=true)[default: (0,0,1.1)]

Returns:

A matrix of production indexes by DMUs and period passages (e.g. "year2 on year1" and "year3 on year2")

Description of the function

prodIndex() displays environmental productivity indices. These productivity measures are implemented for different time periods (eg., years, months etc.) or spatial units (eg., countries, cities etc.), based on the environmental efficiency indicators described in efficiencyScores. Hence, the environmental productivity indices inherit the structure of additive and multiplicative productivity measures.

Interpretation of the results

The additive productivity indicator shows combined desirable and undesirable outputs productivity improvement (respectively decline) when it takes positive (respectively negative) values. In the multiplicative context, if the productivity measure is greater (respectively lesser) than 1 then, desirable and undesirable outputs productivity increase (respectively decrease) arises. The BDisposal package underscores the prominent sources of environmental productivity change. The main drivers of productivity variation are technological change, technical efficiency variation and scale efficiency change. For the additive background, when the technological change is greater (respectively, lesser) than 0 then, technolog- ical improvement (respectively, deterioration) occurs. A similar reasoning applies for the additive technical and scale efficiency components. In the multiplicative context, if the technological change is greater (respectively, lesser) than 1 then, technological in- crease (respectively, decrrease) arises. A similar reasonnng applies for the multiplicative technical and scale efficiency components.

Notes:

The non-convex problem still need to consider bad inputs, consider directions, consider multiplicative/additive structures